Restructure the project, rename the files
