+++ graphic-objects/bullet.h
@@ -2,7 +2,7 @@
 #define BULLET_H
 #include <QGraphicsRectItem>
 #include <QObject>
-#include "Const.h"
+#include "consts.h"
 
 class Bullet: public QObject, public QGraphicsPixmapItem{
     Q_OBJECT
