+++ graphic-objects/hero.h
