+++ graphic-objects/spaceship.h
